User Branch Commit: Completed Zend_Test_PHPUnit_Database
git-svn-id: http://framework.zend.com/svn/framework/standard/branches/user@16463 44c647ce-9c0f-0410-b52a-842ac1e357ba

+++ beberlei/Zend_Entity/library/Zend/Test/PHPUnit/Database/Operation/DeleteAll.php
@@ -2,13 +2,22 @@
 
 class Zend_Test_PHPUnit_Database_Operation_DeleteAll extends PHPUnit_Extensions_Database_Operation_Truncate
 {
+    /**
+     * @param PHPUnit_Extensions_Database_DB_IDatabaseConnection $connection
+     * @param PHPUnit_Extensions_Database_DataSet_IDataSet $dataSet
+     */
     public function execute(PHPUnit_Extensions_Database_DB_IDatabaseConnection $connection, PHPUnit_Extensions_Database_DataSet_IDataSet $dataSet)
     {
+        if(!($connection instanceof Zend_Test_PHPUnit_Database_Connection)) {
+            require_once "Zend/Test/PHPUnit/Database/Exception.php";
+            throw new Zend_Test_PHPUnit_Database_Exception("Not a valid Zend_Test_PHPUnit_Database_Connection instance, ".get_class($connection)." given!");
+        }
+
         foreach ($dataSet as $table) {
             try {
                 $tableName = $table->getTableMetaData()->getTableName();
                 $connection->getConnection()->delete($tableName);
-            } catch (Zend_Db_Exception $e) {
+            } catch (Exception $e) {
                 throw new PHPUnit_Extensions_Database_Operation_Exception('DELETEALL', 'DELETE FROM '.$tableName.'', array(), $table, $e->getMessage());
             }
         }
